♻️ refactor(cli): enhance label command structure
- move Label struct implementation from cli.rs to cli/label.rs
- integrate run_label method within Label struct for better cohesion

✨ feat(cli): improve label command flexibility

- add Parser derive for Label struct to enhance command parsing
- allow optional label customization using author, label, desc, and colour fields
